Authentication Guard

+++ src/routes/posts/posts.controller.ts
@@ -1,36 +1,40 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, UseGuards } from '@nestjs/common'
+import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common'
+import { AuthType, ConditionGuard } from 'src/shared/constants/auth.constant'
+import { ActiveUser } from 'src/shared/decorators/active-user.decorator'
+import { Auth } from 'src/shared/decorators/auth.decorator'
+import { CreatePostBodyDTO, GetPostItemDTO, UpdatePostBodyDTO } from './post.dto'
 import { PostsService } from './posts.service'
-import { APIKeyGuard } from 'src/shared/guards/api-key.guard'
-import { AccessTokenGuard } from 'src/shared/guards/access-token.guard'
 
 @Controller('posts')
 export class PostsController {
   constructor(private readonly postService: PostsService) {}
 
-  @UseGuards(AccessTokenGuard)
-  @UseGuards(APIKeyGuard)
+  @Auth([AuthType.Bearer, AuthType.APIKey], { condition: ConditionGuard.And })
   @Get()
-  getPosts() {
-    return this.postService.getPosts()
+  getPosts(@ActiveUser('userId') userId: number) {
+    return this.postService.getPosts(userId).then((posts) => posts.map((post) => new GetPostItemDTO(post)))
   }
 
   @Post()
-  createPost(@Body() body: any) {
-    return this.postService.createPost(body)
+  @Auth([AuthType.Bearer])
+  async createPost(@Body() body: CreatePostBodyDTO, @ActiveUser('userId') userId: number) {
+    return new GetPostItemDTO(await this.postService.createPost(userId, body))
   }
 
   @Get(':id')
-  getPost(@Param('id') id: string) {
-    return this.postService.getPost(id)
+  async getPost(@Param('id') id: string) {
+    return new GetPostItemDTO(await this.postService.getPost(Number(id)))
   }
 
-  @Put('id')
-  updatePost(@Param('id') id: string, @Body() body: any) {
-    return this.postService.updatePost(id, body)
+  @Put(':id')
+  @Auth([AuthType.Bearer])
+  async updatePost(@Param('id') id: string, @ActiveUser('userId') userId: number, @Body() body: UpdatePostBodyDTO) {
+    return new GetPostItemDTO(await this.postService.updatePost({ postId: Number(id), userId, body }))
   }
 
-  @Delete('id')
-  deletePost(@Param('id') id: string) {
-    return this.postService.deletePost(id)
+  @Delete(':id')
+  @Auth([AuthType.Bearer])
+  deletePost(@Param('id') id: string, @ActiveUser('userId') userId: number): Promise<boolean> {
+    return this.postService.deletePost({ postId: Number(id), userId })
   }
 }
